requirements.txt
requests
beautifulsoup4
uiautomator2
adbutils[apk]
easyocr
Pillow